Make sure that webpack build doesn’t end up requiring `fs`
