+++ lib/browser-index.d.ts
@@ -1,3 +1,4 @@
 export * as keyStores from './key_stores/browser-index';
 export * from './common-index';
+export * from './browser-connect';
 import 'error-polyfill';
