+++ lib/browser-index.js
@@ -24,4 +24,5 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.keyStores = __importStar(require("./key_stores/browser-index"));
 __exportStar(require("./common-index"), exports);
+__exportStar(require("./browser-connect"), exports);
 require("error-polyfill");
