+++ lib/common-index.js
@@ -19,7 +19,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.WalletConnection = exports.WalletAccount = exports.Near = exports.connect = exports.KeyPair = exports.Signer = exports.InMemorySigner = exports.Contract = exports.Connection = exports.Account = exports.multisig = exports.validators = exports.transactions = exports.utils = exports.providers = exports.accountCreator = void 0;
+exports.WalletConnection = exports.WalletAccount = exports.Near = exports.KeyPair = exports.Signer = exports.InMemorySigner = exports.Contract = exports.Connection = exports.Account = exports.multisig = exports.validators = exports.transactions = exports.utils = exports.providers = exports.accountCreator = void 0;
 const providers = __importStar(require("./providers"));
 exports.providers = providers;
 const utils = __importStar(require("./utils"));
@@ -44,7 +44,6 @@ Object.defineProperty(exports, "Contract", { enumerable: true, get: function () 
 const key_pair_1 = require("./utils/key_pair");
 Object.defineProperty(exports, "KeyPair", { enumerable: true, get: function () { return key_pair_1.KeyPair; } });
 const near_1 = require("./near");
-Object.defineProperty(exports, "connect", { enumerable: true, get: function () { return near_1.connect; } });
 Object.defineProperty(exports, "Near", { enumerable: true, get: function () { return near_1.Near; } });
 // TODO: Deprecate and remove WalletAccount
 const wallet_account_1 = require("./wallet-account");
