+++ lib/index.d.ts
@@ -1,2 +1,3 @@
 export * as keyStores from './key_stores/index';
 export * from './common-index';
+export * from './connect';
