+++ lib/index.js
@@ -24,3 +24,4 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.keyStores = __importStar(require("./key_stores/index"));
 __exportStar(require("./common-index"), exports);
+__exportStar(require("./connect"), exports);
