+++ lib/near.js
@@ -3,14 +3,12 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.connect = exports.Near = void 0;
+exports.Near = void 0;
 const bn_js_1 = __importDefault(require("bn.js"));
 const account_1 = require("./account");
 const connection_1 = require("./connection");
 const contract_1 = require("./contract");
-const unencrypted_file_system_keystore_1 = require("./key_stores/unencrypted_file_system_keystore");
 const account_creator_1 = require("./account_creator");
-const key_stores_1 = require("./key_stores");
 class Near {
     constructor(config) {
         this.config = config;
@@ -76,30 +74,3 @@ class Near {
     }
 }
 exports.Near = Near;
-/**
- * Initialize connection to Near network.
- */
-async function connect(config) {
-    // Try to find extra key in `KeyPath` if provided.
-    if (config.keyPath && config.deps && config.deps.keyStore) {
-        try {
-            const accountKeyFile = await unencrypted_file_system_keystore_1.readKeyFile(config.keyPath);
-            if (accountKeyFile[0]) {
-                // TODO: Only load key if network ID matches
-                const keyPair = accountKeyFile[1];
-                const keyPathStore = new key_stores_1.InMemoryKeyStore();
-                await keyPathStore.setKey(config.networkId, accountKeyFile[0], keyPair);
-                if (!config.masterAccount) {
-                    config.masterAccount = accountKeyFile[0];
-                }
-                config.deps.keyStore = new key_stores_1.MergeKeyStore([config.deps.keyStore, keyPathStore]);
-                console.log(`Loaded master account ${accountKeyFile[0]} key from ${config.keyPath} with public key = ${keyPair.getPublicKey()}`);
-            }
-        }
-        catch (error) {
-            console.warn(`Failed to load master account key from ${config.keyPath}: ${error}`);
-        }
-    }
-    return new Near(config);
-}
-exports.connect = connect;
