+++ lib/signer.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.InMemorySigner = exports.Signer = void 0;
 const js_sha256_1 = __importDefault(require("js-sha256"));
 const key_pair_1 = require("./utils/key_pair");
-const key_stores_1 = require("./key_stores");
+const in_memory_key_store_1 = require("./key_stores/in_memory_key_store");
 /**
  * General signing interface, can be used for in memory signing, RPC singing, external wallet, HSM, etc.
  */
@@ -31,7 +31,7 @@ class InMemorySigner extends Signer {
      * @param keyPair The keyPair to use for signing
      */
     static async fromKeyPair(networkId, accountId, keyPair) {
-        const keyStore = new key_stores_1.InMemoryKeyStore();
+        const keyStore = new in_memory_key_store_1.InMemoryKeyStore();
         await keyStore.setKey(networkId, accountId, keyPair);
         return new InMemorySigner(keyStore);
     }
