+++ src/browser-index.ts
@@ -1,4 +1,5 @@
 export * as keyStores from './key_stores/browser-index';
 export * from './common-index';
+export * from './browser-connect';
 
 import 'error-polyfill';
