+++ src/near.ts
@@ -1,15 +1,13 @@
-
 import BN from 'bn.js';
 import { Account } from './account';
 import { Connection } from './connection';
 import { Signer } from './signer';
 import { Contract } from './contract';
-import { readKeyFile } from './key_stores/unencrypted_file_system_keystore';
 import { PublicKey } from './utils/key_pair';
 import { AccountCreator, LocalAccountCreator, UrlAccountCreator } from './account_creator';
-import { InMemoryKeyStore, KeyStore, MergeKeyStore } from './key_stores';
+import { KeyStore } from './key_stores';
 
-type NearConfig = {
+export type NearConfig = {
   keyStore?: KeyStore;
   signer?: Signer;
   deps?: { keyStore: KeyStore };
@@ -90,34 +88,4 @@ export class Near {
         const result = await account.sendMoney(receiver, amount);
         return result.transaction_outcome.id;
     }
-}
-
-type ConnectConfig = NearConfig & {
-  keyPath?: string;
-}
-
-/**
- * Initialize connection to Near network.
- */
-export async function connect(config: ConnectConfig): Promise<Near> {
-    // Try to find extra key in `KeyPath` if provided.
-    if (config.keyPath && config.deps && config.deps.keyStore) {
-        try {
-            const accountKeyFile = await readKeyFile(config.keyPath);
-            if (accountKeyFile[0]) {
-                // TODO: Only load key if network ID matches
-                const keyPair = accountKeyFile[1];
-                const keyPathStore = new InMemoryKeyStore();
-                await keyPathStore.setKey(config.networkId, accountKeyFile[0], keyPair);
-                if (!config.masterAccount) {
-                    config.masterAccount = accountKeyFile[0];
-                }
-                config.deps.keyStore = new MergeKeyStore([config.deps.keyStore, keyPathStore]);
-                console.log(`Loaded master account ${accountKeyFile[0]} key from ${config.keyPath} with public key = ${keyPair.getPublicKey()}`);
-            }
-        } catch (error) {
-            console.warn(`Failed to load master account key from ${config.keyPath}: ${error}`);
-        }
-    }
-    return new Near(config);
 }
